Address latest StateMachine designer review
