+++ src/modules/Elsa.Studio.Workflows/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Elsa.Studio.Workflows.Tests")]
